distributed postprocessing and segmentation
