+++ chunkmydocs/src/utils/workers/mod.rs
@@ -2,4 +2,5 @@ pub mod payload;
 pub mod log;
 pub mod preprocess;
 pub mod segmentation;
+pub mod postprocess;
 pub mod ocr;
