+++ chunkmydocs/src/utils/workers/segmentation.rs
@@ -1,20 +1,16 @@
 use crate::models::rrq::queue::QueuePayload;
-use crate::models::server::extract::{ ExtractionPayload, OcrStrategy };
-use crate::models::server::segment::{Chunk, PdlaSegment, Segment };
+use crate::models::server::extract::ExtractionPayload;
+use crate::models::server::segment::{ PdlaSegment, Segment };
 use crate::models::server::task::Status;
 use crate::utils::configs::extraction_config::Config as ExtractionConfig;
 use crate::utils::configs::task_config::Config as TaskConfig;
 use crate::utils::db::deadpool_postgres::create_pool;
-use crate::utils::json2mkd::json_2_mkd::hierarchical_chunking;
-use crate::utils::services::crop::crop_image;
 use crate::utils::services::pdf::split_pdf;
 use crate::utils::services::pdla::pdla_extraction;
 use crate::utils::storage::config_s3::create_client;
 use crate::utils::storage::services::{ download_to_tempfile, upload_to_s3 };
 use crate::utils::workers::{ log::log_task, payload::produce_extraction_payloads };
 use chrono::Utc;
-use futures::future::try_join_all;
-use rayon::prelude::*;
 use std::io::Write;
 use std::path::PathBuf;
 use tempdir::TempDir;
@@ -112,89 +108,8 @@ pub async fn process(payload: QueuePayload) -> Result<(), Box<dyn std::error::Er
             page_offset += extraction_payload.batch_size.unwrap_or(1) as u32;
         }
 
-        let image_folder_location = extraction_payload.image_folder_location.clone();
-        let page_image_paths = try_join_all(
-            (0..extraction_payload.page_count.unwrap()).map(|i| {
-                let image_folder = image_folder_location.clone();
-                let s3_client = s3_client.clone();
-                let reqwest_client = reqwest_client.clone();
-                async move {
-                    let image_name = format!("page_{}.jpg", i + 1);
-                    let image_location = format!("{}/{}", image_folder, image_name);
-                    download_to_tempfile(&s3_client, &reqwest_client, &image_location, None).await
-                }
-            })
-        ).await?;
-
-        log_task(
-            task_id.clone(),
-            Status::Processing,
-            Some("Cropping segments".to_string()),
-            None,
-            &pg_pool
-        ).await?;
-
-        let cropped_temp_dir = TempDir::new("cropped_images")?;
-        combined_segments.par_iter().for_each(|segment| {
-            let page_index = (segment.page_number as usize) - 1;
-            if let Some(image_path) = page_image_paths.get(page_index) {
-                let crop_result = crop_image(
-                    &image_path.path().to_path_buf(),
-                    segment,
-                    &cropped_temp_dir.path().to_path_buf()
-                ).map_err(|e|
-                    format!(
-                        "Failed to crop image for segment {} on page {}: {:?}",
-                        segment.segment_id,
-                        segment.page_number,
-                        e
-                    )
-                );
-                match crop_result {
-                    Ok(_) => (),
-                    Err(err_msg) => eprintln!("{}", err_msg),
-                }
-            } else {
-                eprintln!("Image not found for page {}", segment.page_number);
-            }
-        });
-
-        for entry in std::fs::read_dir(cropped_temp_dir.path())? {
-            let entry = entry?;
-            let file_name = entry.file_name();
-            let s3_key = format!(
-                "{}/{}",
-                extraction_payload.image_folder_location,
-                file_name.to_string_lossy()
-            );
-            upload_to_s3(&s3_client, &s3_key, &entry.path()).await?;
-            combined_segments
-                .iter_mut()
-                .find(
-                    |segment|
-                        segment.segment_id ==
-                        file_name.clone().to_string_lossy().split(".").next().unwrap()
-                )
-                .map(|segment| {
-                    segment.image = Some(s3_key.clone());
-                });
-        }
-
-        log_task(
-            task_id.clone(),
-            Status::Processing,
-            Some("Chunking segments".to_string()),
-            None,
-            &pg_pool
-        ).await?;
-
-        let chunks: Vec<Chunk> = hierarchical_chunking(
-            combined_segments,
-            extraction_payload.target_chunk_length
-        ).await?;
-
         let mut output_temp_file = NamedTempFile::new()?;
-        output_temp_file.write_all(serde_json::to_string(&chunks)?.as_bytes())?;
+        output_temp_file.write_all(serde_json::to_string(&combined_segments)?.as_bytes())?;
 
         upload_to_s3(
             &s3_client,
@@ -214,25 +129,18 @@ pub async fn process(payload: QueuePayload) -> Result<(), Box<dyn std::error::Er
     match result {
         Ok(_) => {
             println!("Task succeeded");
-            if extraction_payload.configuration.ocr_strategy == OcrStrategy::Off {
-                log_task(
-                    task_id.clone(),
-                    Status::Succeeded,
-                    Some("Task succeeded".to_string()),
-                    Some(Utc::now()),
-                    &pg_pool
-                ).await?;
-            } else {
-                let extraction_config = ExtractionConfig::from_env()?;
-                produce_extraction_payloads(extraction_config.queue_ocr, extraction_payload.clone()).await?;
-                log_task(
-                    task_id.clone(),
-                    Status::Processing,
-                    Some("OCR queued".to_string()),
-                    None,
-                    &pg_pool
-                ).await?;
-            }
+            let extraction_config = ExtractionConfig::from_env()?;
+            produce_extraction_payloads(
+                extraction_config.queue_ocr,
+                extraction_payload.clone()
+            ).await?;
+            log_task(
+                task_id.clone(),
+                Status::Processing,
+                Some("Chunking queued".to_string()),
+                None,
+                &pg_pool
+            ).await?;
             Ok(())
         }
         Err(e) => {
